Move code into .js files per obj/add related test.js
