+++ lib/car.js
@@ -34,7 +34,4 @@ Car.prototype.move = function() {
   }
 };
 
-
-
-
 module.exports = Car;
